fix(daemon): apply round-2 codex review findings to #717 fix
Round-2 adversarial review surfaced four findings worth addressing in
this PR (the rest tracked as separate issues).

Defense #1 (medium) — buildProjectionNormalized's projection==nil
branch still passed result.Status through verbatim, so any future
caller that forgets to set Status would re-trigger the same
empty-status leak. Add a fail-safe default in the helper: if
projection==nil and Status is empty, default to StatusClear. This is
belt-and-suspenders alongside the explicit StatusClear at sweep.go's
afterFrameCleared callsite — both layers now agree.

Attack #1 (medium) — race window between projectionForSession and
Broadcast in afterFrameCleared. A hook handler could create a new
frame for the same tmux session in that window (e.g. user kills
opencode and immediately runs `opencode` again, SessionStart hook
lands mid-sweep). The stale broadcast would then overwrite the newly
installed running status with clear. Re-resolve projection right
before broadcasting and use the fresh value, so the broadcast carries
the live state. Race is best-effort; full serialization requires
per-session locking which is tracked separately.

Health #1 + #2 (medium) — sweep_test.go was already 2815 lines mixing
many sweep concerns. Move the two new broadcast contract tests to a
focused sweep_broadcast_test.go, and extract a readSweepNormalizedEvent
helper that decodes the HostEvent envelope and inner NormalizedEvent
once. Future sweep broadcast tests now have an obvious home and a
consistent assertion pattern.
